Handle commands and Arguments of Type String and Int
This is the start of the library. It implements several attributes
as well as the reflection code necessary to work with classes
decorated with those attributes.

At this point the library can do the following:
 - Find all classes designated as CommandClass
 - Find all methods of those classes that are designated as CommandHandler
 - Based on the program arguments, find an instantiate the appropriate
CommandClass
 - Based on the number of parameters, find the appropriate CommandHandler
method and invoke the method, passing the parameters.
 - if the parameter is not a string, the system will attempt to convert the
type IF the parameter is of type int