Use a Func For Type Conversion Instead of Interface
Using an interface obligates users of this library to implement
a class to handle type conversion. After some use and experimentation
a lot of these classes are simply implementing [type].Parse(string).
A class is overkill for the amount of code required to perform the
conversion. Using a Func<string, object> provides the same functionality
in a much more compact and concise format.

Resolves: #5