Update Option Parsing Logic
This commit updates the option parsing logic to be much more versatile and
compatible with Git's handling of options. Aggregate options can now have 1
option included that takes a value; Abreviated options can now take arguments
and argument values can either be designated with an '=' or will be assumed
to be the next argument supplied.