Update Handler Filtering to Accept Arrays and Overloads
This update allows the system to handle MethodHandlers with array inputs. If a method handler will have an array argument it can only have one and that array must be the only argument for the method signature.

In addition to being able to handle arrays as inputs, the system will now handle overloaded method handlers. Handlers can now differ in their argument types and the system will try each one. CLImber consideres arguments that are NOT strings to be more "specific" than a string argument. This is how overloaded methods are prioritized. In other words, if two methods each have 5 parameters but one is all strings and the other is 3 integers, CLImber will prefer the method with 3 integers. If that method call fails it will try the method that takes all string arguments.Release 0.6.0
This release allows commands to be configured as case sensitive as an property of a CLImber
object. It also updates the way options are handled to be much closer to the way Git handles
them.

Resolves: #18